rm uint wrapper

+++ packages/revolution/src/AuctionHouse.sol
@@ -405,7 +405,7 @@ contract AuctionHouse is
         if (block.timestamp < _auction.endTime) revert AUCTION_NOT_COMPLETED();
 
         auction.settled = true;
-        auctions[_auction.tokenId].settledBlockWad = uint256(block.number) * 1e18;
+        auctions[_auction.tokenId].settledBlockWad = block.number * 1e18;
 
         PaidToCreators memory paidToCreators = PaidToCreators({ eth: 0, points: 0 });
 
@@ -433,7 +433,7 @@ contract AuctionHouse is
                     amount: _auction.amount,
                     winner: _auction.bidder,
                     amountPaidToOwner: 0,
-                    settledBlockWad: uint256(block.number) * 1e18
+                    settledBlockWad: block.number * 1e18
                 });
             }
